Added possibility of regrow bloons and made it so that bloons can actually get pierced instead of getting hit multipe times. they can still sort of get hit multiple times but not from the same projectile or at least not as easily so imo good enough

+++ src/main/java/Game/WorldStuff/TdWorld.java
@@ -281,7 +281,7 @@ public class TdWorld implements World {
     return health;
   }
 
-  public void changeHealth(int change) {
+  public void changeHealth(float change) {
     health += change;
     resourceTracker.update();
   }
